Refactor pow benchmark to its own file
